销售模块提交

+++ ruoyi-system/src/main/java/com/ruoyi/system/mapper/GsGoodsSnMapper.java
@@ -33,7 +33,7 @@ public interface GsGoodsSnMapper {
 
     int updateByPrimaryKey(GsGoodsSn record);
 
-    List<GsGoodsSn> selectOutByWhIdAndGoodsId(@Param("whId") Integer whId, @Param("goodsId") Integer goodsId, @Param("lockQty") Double lockQty);
+    List<GsGoodsSn> selectOutByWhIdAndGoodsId(@Param("whId") Integer whId, @Param("goodsId") Integer goodsId, @Param("lockQty") Integer lockQty);
 
     List<TakeGoodsOrderListVo> selectSwJsTaskGoodsRelLists(TakeGoodsOrderListDto takeGoodsOrderListDto);
 }
